requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.2.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
